Melhor implementacao do algoritmo ratemonotonic com suporte a preempcao e sem preempcao, ainda ficando pendente de melhorias futuras. para esse mecanismo foram gerados 10 testes unitarios que validam com preempcao e sem preempcao.

+++ source/scheduler/src/main/java/br/eti/hussamaismail/scheduler/domain/RateMonotonicScheduler.java
@@ -20,7 +20,7 @@ public class RateMonotonicScheduler extends MonotonicScheduler {
 	public Chart simulate() throws DeadlineNotSatisfiedException {
 		this.getTasksUtil().sortTasksByPeriod(this);
 		super.simulate();
-		double higherPeriod = getTasksUtil().getHigherPeriodFromPeriodicTasks(getTasks());
+		int higherPeriod = getTasksUtil().getHigherPeriodFromPeriodicTasks(getTasks());
 		Map<Integer, List<PeriodicTask>> mapWithPeriodsAndTasks = getTasksUtil().getMapWithPeriodsAndTasks(getTasks());			
 		
 		return generateMonotonicChart(mapWithPeriodsAndTasks, higherPeriod);
